requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
distro==1.9.0
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.24.0
fsspec==2026.2.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pillow==12.1.1
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
scikit-image==0.26.0
scipy==1.17.0
setuptools==82.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tifffile==2026.1.28
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
